+++ CHANGES.md
@@ -24,6 +24,7 @@ Apollo 2.5.0
 * [Fix: Operate the AccessKey multiple times within one second](https://github.com/apolloconfig/apollo/pull/5490)
 * [Bugfix: Prevent accidental cache deletion when recreating AppNamespace with the same name and appid](https://github.com/apolloconfig/apollo/issues/5502)
 * [Feature: Support ordinary users to modify personal information](https://github.com/apolloconfig/apollo/pull/5511)
+* [Feature: Support exporting and importing configurations for specified applications and clusters](https://github.com/apolloconfig/apollo/pull/5517)
 * [doc: Add rust apollo client link](https://github.com/apolloconfig/apollo/pull/5514)
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo/milestone/16?closed=1)